Switch of special page structure from json key, value to array of key value.  Fix for deletion of main survey start with no image/video.

+++ src/TRAISI/ClientApp/src/admin-app/app/survey-builder/components/special-page-builder/main-survey-access1/main-survey-access1.component.ts
@@ -142,10 +142,12 @@ export class MainSurveyAccess1Component implements OnInit {
 		} else if (this.imageSource) {
 			uploadPath = new UploadPath(this.imageSource);
 		}
-		this.surveyBuilderService.deleteUploadedFile(uploadPath).subscribe();
-		this.videoSource = undefined;
-		this.imageSource = undefined;
-		this.updateImageVideoContent();
+		if (uploadPath) {
+			this.surveyBuilderService.deleteUploadedFile(uploadPath).subscribe();
+			this.videoSource = undefined;
+			this.imageSource = undefined;
+			this.updateImageVideoContent();
+		}
 	}
 
 	clearUploads() {
